src/api-service/pyproject.toml
[project]
name = "tummyai-app-api"
version = "0.1.0"
description = "TummyAI App API Service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=24.1.1",
    "fastapi>=0.111.0",
    "flake8>=7.0.0",
    "httpx>=0.26.0",
    "pre-commit>=3.6.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest>=7.4.4",
    "python-multipart>=0.0.7",
    "requests>=2.31.0",
    "uvicorn>=0.27.0",
    "chromadb>=1.3.0",
    "fastmcp>=2.13.0.2",
    "google-genai>=1.47.0",
    "pillow>=12.0.0",
    "tensorflow==2.19.0",
    "pandas>=2.3.3",
    "numpy>=2.1.3",
    "scipy>=1.15.3",
    "statsmodels==0.14.5",
    "google-cloud>=0.34.0",
    "google-cloud-storage>=3.6.0",
]
